# openapi.json
{
...
  "info": {
    "description": "",
    "license": {
      "name": ""
    },
    "title": "adapters",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/assets/{file}": {
      "get": {
        "operationId": "serve_asset",
        "parameters": [
          {
            "description": "`{job_id}.svg`",
            "in": "path",
            "name": "file",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "finalized artifact bytes"
          },
          "404": {
            "description": "unknown job or artifact"
          },
          "409": {
            "description": "job not finalized"
          }
        },
        "tags": [
          "video"
        ]
      }
    },
...
    "/users/{user_id}/share_card/{market_id}": {
      "get": {
        "operationId": "share_card",
        "parameters": [
          {
            "description": "user",
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "market",
            "in": "path",
            "name": "market_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "share-card SVG (post-resolution only)"
          },
          "404": {
            "description": "unknown market/user or no participation"
          },
          "409": {
            "description": "market not resolved yet"
          }
        },
        "tags": [
          "video"
        ]
      }
    },
...
  },
...
}